+++ Kernel/System/DynamicField/Driver/TextArea.pm
@@ -383,46 +383,6 @@ EOF
     return $Data;
 }
 
-sub SearchFieldParameterBuild {
-    my ( $Self, %Param ) = @_;
-
-    # get field value
-    my $Value = $Self->SearchFieldValueGet(%Param);
-
-    if ( !$Value ) {
-        return {
-            Parameter => {
-                'Like' => '',
-            },
-            Display => '',
-            }
-    }
-
-    # return search parameter structure
-    return {
-        Parameter => {
-            'Like' => '*' . $Value . '*',
-        },
-        Display => $Value,
-    };
-}
-
-sub StatsSearchFieldParameterBuild {
-    my ( $Self, %Param ) = @_;
-
-    my $Value = $Param{Value};
-
-    if ( !$Value ) {
-        return {
-            'Like' => '',
-        };
-    }
-
-    return {
-        'Like' => '*' . $Value . '*',
-    };
-}
-
 1;
 
 =head1 TERMS AND CONDITIONS